# Added performance enhancements (XLA, AMP) to examples