@thomwolf 2019-10-15
Great, looks good to me cc @LysandreJik.

Adding the info of the PR in the example README.